requirements.txt
langchain
langchain-openai
langchain-community
openai
pypdf
PyPDF2
python-dotenv
tiktoken
pdf2image
easyocr
Pillow
fastapi
uvicorn
pinecone-client
langchain-pinecone